adicionando os primeiro vídeos ao site
