+++ src/components/IndexPage/index.tsx
@@ -1,0 +1,11 @@
+import GeneralRemunerationSection from './GeneralRemunerationSection';
+import HeadlineSection from './HeadlineSection';
+import TransparencySection from './TransparencySection';
+
+const IndexPage = {
+  HeadlineSection,
+  TransparencySection,
+  GeneralRemunerationSection,
+};
+
+export default IndexPage;
