+++ src/components/Video/index.tsx
@@ -1,0 +1,7 @@
+import Player from './Player';
+
+const Video = {
+  Player,
+};
+
+export default Video;
